requirements.txt
anthropic>=0.15.0
PyYAML>=6.0
PyPDF2>=3.0.0
requests>=2.25.1
feedparser>=6.0.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
arxiv>=1.4.7
pandas>=2.0.0
numpy>=1.20.0
matplotlib>=3.5.0